Improve iOS Simulator compatibility
`PanoramaView`, `StereoView` and `StereoViewController` now can be used
on iOS simulator.

```
#if METALSCOPE_ENABLE_METAL
let panoramaView = PanoramaView(frame: view.bounds, device: device)
#else
let panoramaView = PanoramaView(frame: view.bounds) // for Simulator
#endif
```

Please note that these classes are significantly limited in
functionality on the simulator. For example, `PanoramaView` can display
photos, but cannot display videos. For `StereoView` and
`StereoViewController`, it is a placeholder on the simulator and
nothing is displayed.